% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersLoupe.R
\name{compute_intensities}
\alias{compute_intensities}
\title{Compute instesities same way as loupe browser}
\usage{
compute_intensities(x, cb_a, cb_b)
}
\arguments{
\item{x}{\itemize{
\item Sparse matrix of counts (feature x cell).
}}

\item{cb_a}{\itemize{
\item Vector of indices of cells in group A
}}

\item{cb_b}{\itemize{
\item Vector of indices of cells in group B
}}
}
\value{
data.frame with 3 columns: gname, id.1.intensity - intensity of specific
gene in group A, id.2.intensity of specific gene in group B
}
\description{
Compute instesities same way as loupe browser
}
