% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersLoupe.R
\name{compute_sseq_params}
\alias{compute_sseq_params}
\title{Compute global parameters for the sSeq differential expression method}
\usage{
compute_sseq_params(x, zeta_quantile = SSEQ_ZETA_QUANTILE)
}
\arguments{
\item{x}{Sparse matrix of counts (feature x cell)}

\item{zeta_quantile}{(float) - Quantile of method-of-moments
dispersion estimates to use as the shrinkage target zeta.

The key parameters are the shrunken feature-wise dispersions.

This method was published in: Yu D, et al. (2013) Shrinkage estimation of
dispersion in Negative Binomial models for RNA-seq experiments with small
sample size.
Bioinformatics. 29: 1275-1282. doi: 10.1093/bioinformatics/btt143
Exported from cellranger sources:
https://github.com/10XGenomics/cellranger/blob/master/lib/python/cellranger/analysis/diffexp.py}
}
\value{
A list containing the sSeq parameters and some diagnostic info.
}
\description{
Compute global parameters for the sSeq differential expression method
}
