% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersLoupe.R
\name{estimate_size_factors}
\alias{estimate_size_factors}
\title{Estimate size factors (related to cell RNA content and GEM-to-GEM technical variance)}
\usage{
estimate_size_factors(x)
}
\arguments{
\item{x}{Sparse matrix of counts (feature x cell)}
}
\value{
Array of floats, one per cell
}
\description{
Estimate size factors (related to cell RNA content and GEM-to-GEM technical variance)
}
