% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersLoupe.R
\name{nb_asymptotic_test}
\alias{nb_asymptotic_test}
\title{Compute p-value for a pairwise exact test using a fast beta approximation
to the conditional joint distribution of (x_a, x_b).}
\usage{
nb_asymptotic_test(x_a, x_b, size_factor_a, size_factor_b, mu, phi = 0)
}
\arguments{
\item{x_a}{\itemize{
\item (vector) Total count for a single feature in group A
}}

\item{x_b}{\itemize{
\item (vector) Total count for a single feature in group B
}}

\item{size_factor_a}{\itemize{
\item (float) Sum of size factors for group A
}}

\item{size_factor_b}{\itemize{
\item (float) Sum of size factors for group B
}}

\item{mu}{\itemize{
\item (vector) Common mean count for this feature
}}

\item{phi}{\itemize{
\item (vector) Common dispersion for this feature
}}
}
\value{
-p-value the probability that a random pair of counts under
the null hypothesis is more extreme than the observed pair of counts.
}
\description{
Robinson MD and Smyth GK (2008). Small-sample estimation of negative binomial
dispersion, with applications to SAGE data. Biostatistics, 9, 321-332
It is based a method-of-moments gamma approximation to the negative binomial distribution.
}
\details{
Adapted from implementation in the "edgeR" package:
https://rdrr.io/bioc/edgeR/src/R/exactTestBetaApprox.R

Added small adjustments related to size_factors which should be uniform with
similar function from cellranger sources
}
