% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersLoupe.R
\name{nb_exact_test}
\alias{nb_exact_test}
\title{Compute p-value for a pairwise exact test using the negative binomial.
Extracted from https://github.com/hb-gitified/cellrangerRkit}
\usage{
nb_exact_test(x_a, x_b, s_a, s_b, mu, phi)
}
\arguments{
\item{x_a}{\itemize{
\item (int) Total count for a single feature in group A
}}

\item{x_b}{\itemize{
\item (int) Total count for a single feature in group B
}}

\item{s_a}{\itemize{
\item (float) Sum of size factors for group A
}}

\item{s_b}{\itemize{
\item (float) Sum of size factors for group B
}}

\item{mu}{\itemize{
\item (float) Common mean count for this feature
}}

\item{phi}{\itemize{
\item (float) Common dispersion for this feature
}}
}
\value{
p-value (float); the probability that a random pair of counts under
the null hypothesis is more extreme than the observed pair of counts
}
\description{
Compute p-value for a pairwise exact test using the negative binomial.
Extracted from https://github.com/hb-gitified/cellrangerRkit
}
