% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersLoupe.R
\name{sseq_differential_expression}
\alias{sseq_differential_expression}
\title{sSeq pairwise differential expression test.}
\usage{
sseq_differential_expression(x, cond_a, cond_b, sseq_params, big_count = 900)
}
\arguments{
\item{x}{\itemize{
\item Sparse matrix of counts (feature x cell).
}}

\item{cond_a}{\itemize{
\item Vector of indices of cells in group A
}}

\item{cond_b}{\itemize{
\item Vector of indices of cells in group B
}}

\item{sseq_params}{\itemize{
\item Precomputed global parameters
}}

\item{big_count}{\itemize{
\item Use asymptotic approximation if both counts > this
}}
}
\value{
A data.frame with DE results for group A relative to group B
}
\description{
Exported from cellranger sources.
https://github.com/10XGenomics/cellranger/blob/master/lib/python/cellranger/analysis/diffexp.py
}
